chore: rename inline directive
