+++ src/core/parser/markdown/directive/micromark-directive/tokenize-directive-inline.ts
